01_batch_etl_pipeline/airflow_dag.py
# Example Airflow DAG skeleton

from airflow import DAG
from airflow.operators.python import PythonOperator
from datetime import datetime

# TODO: Define ETL tasks here

def dummy_task():
    print("ETL Task executed")

dag = DAG('batch_etl_pipeline', start_date=datetime(2023,1,1), schedule_interval='@daily')
